requirements.txt
# Core ML/DL
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
facenet-pytorch==2.5.3

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1
scikit-image==0.22.0

# Data Science
numpy==1.24.4
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# API & Serving
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
aiofiles==23.2.1

# UI
streamlit==1.28.2

# Experiment Tracking
tensorboard==2.15.1
tqdm==4.66.1

# Config & Utils
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
